requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
azure-keyvault-secrets
azure-identity
azure-storage-queue
azure-storage-blob
azure-data-tables
uuid
pdfminer.six
python-docx
openai
nltk
tiktoken
tenacity
pytest
pandas